requirements.txt
fastapi
uvicorn
requests
openai-whisper
librosa
numpy
matplotlib
sqlite-utils
python-multipart